chore(hype-ui): Renamed api folder for pools
